+++ changelog.md
@@ -3,3 +3,8 @@
 ## 1.0.0
 
 * First initial release
+
+## 1.0.1
+
+* Automatically find relevent images for a GroupedProduct
+* Automatically merge all images from a group